Extract shared client helpers (currency, cost, clipboard, code fence)
- Move duplicated formatCurrency + parseCost out of the two dashboard cards
  (-DashboardCoursesByAmortization, -DashboardUnderutilizedProviders) into utils/
- Move duplicated stripCodeFence and the clipboard-with-textarea-fallback helper
  (copyTextToClipboard) out of ModuleSuggestDialog and BlipLlmAssist into utils/;
  BlipLlmAssist's copyPrompt now wraps the shared helper
