+++ packages/client/src/utils/copyTextToClipboard.ts
@@ -1,0 +1,32 @@
+// Copy text to the clipboard, falling back to a hidden textarea + execCommand
+// when the async Clipboard API is unavailable or denied. Returns whether the
+// copy succeeded.
+export async function copyTextToClipboard(text: string): Promise<boolean> {
+  if (navigator.clipboard?.writeText) {
+    try {
+      await navigator.clipboard.writeText(text);
+      return true;
+    }
+    catch {
+      // fall through to the textarea fallback
+    }
+  }
+  try {
+    const textarea = document.createElement("textarea");
+    textarea.value = text;
+    textarea.setAttribute("readonly", "");
+    textarea.style.position = "fixed";
+    textarea.style.top = "0";
+    textarea.style.left = "0";
+    textarea.style.opacity = "0";
+    document.body.appendChild(textarea);
+    textarea.focus();
+    textarea.select();
+    const ok = document.execCommand("copy");
+    document.body.removeChild(textarea);
+    return ok;
+  }
+  catch {
+    return false;
+  }
+}
