+++ packages/client/src/utils/formatCurrency.ts
@@ -1,0 +1,7 @@
+// Format a number as USD currency.
+export function formatCurrency(value: number): string {
+  return value.toLocaleString("en-US", {
+    style: "currency",
+    currency: "USD",
+  });
+}
