+++ packages/client/src/utils/index.ts
@@ -1,8 +1,13 @@
+/* eslint-disable import/max-dependencies */
+export * from "./copyTextToClipboard";
 export * from "./dailyHelpers";
 export * from "./fetchFunctions";
 export * from "./fieldContext";
+export * from "./formatCurrency";
 export * from "./formHasChanges";
 export * from "./isHttpUrl";
 export * from "./makePercentageComplete";
+export * from "./parseCost";
+export * from "./stripCodeFence";
 export * from "./useIsFieldInvalid";
 export * from "./uuid";
